Changes for Implementaion of Unit of work , optimizaiotn of queries to the Db, global confirmation service to maintain single desing of modal

+++ client/src/app/messages/messages.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -16,7 +17,7 @@ pageNumber = 1;
 pageSize = 5;
 loading = false;
 
-  constructor(private messageService : MessageService) { }
+  constructor(private messageService : MessageService, private confrimService : ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -39,8 +40,13 @@ loading = false;
   }
 
   deleteMessage(id : number){
-    this.messageService.deleteMessage(id).subscribe(()=>{
-      this.messages.splice(this.messages.findIndex(m =>m.id === id), 1);
+
+    this.confrimService.confirm('Confirm delege message', 'This cannot be undone').subscribe(result =>{
+      if(result){
+        this.messageService.deleteMessage(id).subscribe(()=>{
+          this.messages.splice(this.messages.findIndex(m =>m.id === id), 1);
+        })
+      }
     })
   }
 }
